New models

+++ mecanicaRubio/Models/ModelCrypto.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 struct ModelCrypto : Codable {
-    let current_total : Float
-    let purchase_total : Float
+    let current_total : Double
+    let purchase_total : Double
     let data : [Investment]
 }
 
@@ -11,8 +11,8 @@ struct Investment : Codable, Identifiable {
     let book : String
     let amount : String
     let price : String
-    let purchase_value : Float
-    let current_value : Float
+    let purchase_value : Double
+    let current_value : Double
     let updated_at : Date
     let created_at : Date
 }
